feat(DriveTrain): функции для управления скоростей в нормальных системах координат

+++ include/Voltmeter.h
@@ -6,7 +6,7 @@
 float gVoltmeterVolts = 1.0;
 
 float decodeVoltmeter(int rawVolt){
-    return rawVolt * ANALOG_TO_VOLTS;
+    return (float)rawVolt * ANALOG_TO_VOLTS;
 }
 
 void voltInit(){
